+++ examples/modules/golden-signal-alerts-new/variables.tf
@@ -2,7 +2,7 @@ variable "service" {
   description = "The service to create alerts for"
   type = object({
     name                       = string
-    duration                   = number
+    threshold_duration         = number
     cpu_threshold              = number
     response_time_threshold    = number
     error_percentage_threshold = number
@@ -10,7 +10,7 @@ variable "service" {
   })
 }
 
-variable "alert_channel_ids" {
-  description = "The notification channel IDs to add to this policy"
-  type        = list(number)
+variable "notification_channel_ids" {
+  description = "The IDs of notification channels to add to this policy"
+  type        = list(string)
 }